+++ packages/create-sitecore-jss/src/common/processes/transform.ts
@@ -227,12 +227,12 @@ type TransformOptions = {
 
 /**
  * Handles each template file and applies ejs renderer, also:
- * * determines files for copy
- * * determines files for skip
- * * if some files already exist:
- *   * merges package.json files
- *   * concatenates .env files
- *   * compares diffs
+ * - Determines files for copy.
+ * - Determines files for skip.
+ * if some files already exist:
+ *   - merges package.json files
+ *   - concatenates .env files
+ *   - compares diffs
  * @param {string} templatePath path to the template
  * @param {BaseArgs} answers CLI arguments
  * @param {TransformOptions} options custom options