+++ packages/create-sitecore-jss/src/templates/nextjs/scripts/generate-component-builder/index.ts
@@ -1,4 +1,4 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
+// eslint-disable-next-line @typescript-eslint/no-require-imports
 const plugins = require('scripts/temp/generate-component-builder-plugins');
 import { PackageDefinition, ComponentFile } from '@sitecore-jss/sitecore-jss-dev-tools';